requirements.txt
discord.py
feedparser
python-dotenv
google-generativeai